acpi: madt: Silence compiler warning when using Clang
Clang complains:

acpi/madt/madt.c:507:8: error: format specifies type 'unsigned char' but the
 argument has type 'unsigned short' [-Werror,-Wformat]
  500 |    fwts_failed(fw, LOG_LEVEL_MEDIUM,
      |    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  501 |                "SPECMADTFADTRevisions",
      |                ~~~~~~~~~~~~~~~~~~~~~~~~
  502 |                "MADT revision is not in sync with "
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  503 |                "the FADT revision;\n"
      |                ~~~~~~~~~~~~~~~~~~~~~~
  504 |                "MADT %" PRIu8 " expects FADT %" PRIu8 ".%" PRIu8 " "
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  505 |                "but found %" PRIu8 ".%" PRIu8 " instead.",
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  506 |                madt->header.revision,
      |                ~~~~~~~~~~~~~~~~~~~~~~
  507 |                ms->major_version, ms->minor_version,
      |                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  508 |                fadt_major, fadt_minor);
      |                ~~~~~~~~~~~~~~~~~~~~~~~
acpi/madt/madt.c:507:27: error: format specifies type 'unsigned char' but the
 argument has type 'unsigned short' [-Werror,-Wformat]
  500 |    fwts_failed(fw, LOG_LEVEL_MEDIUM,
      |    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  501 |                "SPECMADTFADTRevisions",
      |                ~~~~~~~~~~~~~~~~~~~~~~~~
  502 |                "MADT revision is not in sync with "
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  503 |                "the FADT revision;\n"
      |                ~~~~~~~~~~~~~~~~~~~~~~
  504 |                "MADT %" PRIu8 " expects FADT %" PRIu8 ".%" PRIu8 " "
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  505 |                "but found %" PRIu8 ".%" PRIu8 " instead.",
      |                ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  506 |                madt->header.revision,
      |                ~~~~~~~~~~~~~~~~~~~~~~
  507 |                ms->major_version, ms->minor_version,
      |                ~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~
  508 |                fadt_major, fadt_minor);
      |                ~~~~~~~~~~~~~~~~~~~~~~~

Either the format string should be fixed or the type of the major_version
and the  minor_version field should be adjusted. Looking at the code,
the major_version and minor_version fields are used for comparisons
with the fadt_major and fadt_minor variables, and those are declared
as uint8_t. So let's change the major_version and minor_version fields
to uint8_t, too, so that the involved values have all the same type, which
also matches the format string and thus fixes the complaints from Clang.

Acked-by: Ivan Hu <ivan.hu@canonical.com>

+++ src/acpi/madt/madt.c
@@ -128,8 +128,8 @@
 #define SBBR_ACPI_MAJOR_VERSION 6
 
 struct acpi_madt_subtable_lengths {
-	unsigned short major_version;	/* from revision in FADT header */
-	unsigned short minor_version;	/* FADT field starting with 5.1 */
+	uint8_t major_version;		/* from revision in FADT header */
+	uint8_t minor_version;		/* FADT field starting with 5.1 */
 	unsigned short madt_version;	/* MADT revision */
 	unsigned short num_types;	/* types possible for this version */
 	unsigned short lengths[NUM_SUBTABLE_TYPES];
